% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifalsnm.R
\name{ZIFA_LSNM}
\alias{ZIFA_LSNM}
\title{Zero Inflated Factor Analysis Logistic Skew Normal Multinomial (ZIFA - LSNM) model}
\usage{
ZIFA_LSNM(
  X,
  num_fac = 2,
  NU1P = 1,
  NU2P = 3,
  G1P = 200,
  G2P = 1/200,
  AlphaP = 2
)
}
\arguments{
\item{X}{Count matrix of samples and taxa.}

\item{num_fac}{Number of latent factors.}

\item{NU1P}{Prior value for the shape parameter of \eqn{\kappa_j}.}

\item{NU2P}{Prior value for the shape parameter of \eqn{\kappa_j}.}

\item{G1P}{Prior value for the shape parameter of \eqn{\delta_{jt}}.}

\item{G2P}{Prior value for the rate parameter of \eqn{\delta_{jt}}.}

\item{AlphaP}{Prior value for the skewness parameter of latent factors \eqn{F_{it}}.}
}
\value{
Returns optimized values
}
\description{
ZIFA-LSNM model's objective is to achieve effective dimension
reduction to manage high dimensionality, accounting for the microbiome
data compositional nature, zero inflation and, it utilizes
skew-normal priors on the latent factors to explicitly model skewness
in log-ratio transformation.
}
\examples{
\donttest{
N <- 50
P <- 50
K <- 2

alpha_prior <- 2
alpha_star <- -sqrt(2 / pi) * (alpha_prior / sqrt(1 + alpha_prior^2))
f <- matrix(sn::rsn(N * K, alpha_star, 1, alpha_prior), N, K, TRUE)

beta <- matrix(runif(P * K, -2.5, 2.5), nrow = P, ncol = K, byrow = TRUE)

kappa <- runif(P)

beta_0 <- runif(P, -2, 2)

z <- matrix(0, N, P, byrow = TRUE)
for (i in 1:N) {
  z[i, ] <- rbinom(P, size = 1, prob = kappa)
}
linear_predictor <- matrix(beta_0, N, P, byrow = TRUE) + tcrossprod(f, beta)
rho <- (1 - z) * exp(linear_predictor) / (rowSums((1 - z) * exp(linear_predictor)) + 1e-8)

cmat <- matrix(0, N, P)

for (i in 1:N) cmat[i, ] <- rmultinom(1, as.integer(runif(1, 80000, 100000)), prob = rho[i, ])

OUTPUT <- ZIFA_LSNM(X = cmat, num_fac = K, NU1P = 1, NU2P = 3, G1P = 200, G2P = 1 / 200, AlphaP = 2)
}

}
